Moved resumeRunningStatesFromAnchorRequestStore(..) to a separate service class
