+++ packages/core/src/index.ts
@@ -2,7 +2,7 @@ import * as PubsubMessage from './pubsub/pubsub-message.js'
 
 export { PubsubMessage }
 export { Ceramic } from './ceramic.js'
-export { Repository } from './state-management/repository.js'
+export { AnchorResumingService } from './state-management/anchor-resuming-service.js'
 
 export * from './store/ikv-store.js'
 export * from './ceramic.js'
